프로젝트 정리

+++ backend/src/main/java/com/inspire12/homepage/repository/FileMetaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FileMetaRepository extends JpaRepository<FileMeta, Integer> {
+public interface FileMetaRepository extends JpaRepository<FileMeta, Long> {
 }
